Desktop is nearly done
Finished the layout and started to work on the windows. First one is
almost done. It's still necessary to make it editable.
Also, we should try and make the code I have for the window more easily
reusable. Maybe stuff it into a variable or make a class out of it.